feat(isDirty): change isDirty from a vector to just a boolean

+++ tests/geometry/test_mesh.cpp
@@ -6,9 +6,9 @@ namespace butter {
 TEST(Mesh, CreateTriangle){
     Mesh mesh;
 
-    size_t p0 = mesh.addPoint(1.0, 5.0, 2.0);
-    size_t p1 = mesh.addPoint(2.0, 5.0, -8.0);
-    size_t p2 = mesh.addPoint(-1.0, 5.0, 2.0);
+    uint32_t p0 = mesh.addPoint(1.0, 5.0, 2.0);
+    uint32_t p1 = mesh.addPoint(2.0, 5.0, -8.0);
+    uint32_t p2 = mesh.addPoint(-1.0, 5.0, 2.0);
 
     EXPECT_EQ(mesh.points.size(), 3);
     EXPECT_EQ(mesh.primitives.size(), 0);
